feat: Add equation numbering filter and support for additional equation environments
The code changes in this commit add support for equation numbering in additional equation environments such as "equation*", "align", and "subequations". This is achieved by modifying the "add_eqn_no" function to check for these environments and append the equation number accordingly.

This commit also includes some cleanup and refactoring.